Add files via upload
This is the first draft for our UML diagram. Please look it over and make sure it's right. I probably forgot something or used the wrong arrows. Feel free to make changes as you see fit, but please let everyone know what you change so we're on the same page.
-Christina